+++ src/classes/cells.py
@@ -1,6 +1,6 @@
 import json
 
-from src.classes.actions import ActionSet
+from classes.actions import ActionSet
 
 with open('src/dicts/controls.json', 'r') as file:
     controls = json.load(file)
