+++ src/classes/players.py
@@ -1,7 +1,7 @@
 import json
 import random
 
-from src.classes.player_inventory import PlayerInventory
+from classes.player_inventory import PlayerInventory
 
 with open('src/dicts/attributes.json', 'r') as file:
     attributes = json.load(file)
