+++ src/main.py
@@ -1,4 +1,4 @@
-from src.classes.interface import Interface
+from classes.interface import Interface
 
 game = Interface()
 game.generate_map(rows=5,
